requirements.txt
openai
requests
tiktokapi
flask
python-dotenv